Flatten submodule structure
