+++ example/db.ts
@@ -1,0 +1,4 @@
+import { PrismaClient } from '@prisma/client';
+import prismaRandom from '../dist';
+
+export const prisma = new PrismaClient().$extends(prismaRandom());
