+++ example/index.ts
@@ -1,4 +1,20 @@
-import { PrismaClient } from '@prisma/client';
-import prismaRandom from '../dist';
+import { prisma } from './db';
 
-export const prisma = new PrismaClient().$extends(prismaRandom());
+const main = async () => {
+  const user = await prisma.user.findRandom();
+
+  const post = await prisma.post.findManyRandom(10, {
+    select: { id: true, title: true },
+    where: {
+      OR: [
+        { title: { contains: 'prisma' } },
+        { content: { contains: 'prisma' } },
+      ],
+      published: true,
+    },
+  });
+
+  console.log({ user, post });
+};
+
+main();
