--- example/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "example",
-  "scripts": {
-    "dev": "tsx prisma/test.ts"
-  },
-  "prisma": {
-    "seed": "tsx prisma/seed.ts"
-  },
-  "dependencies": {
-    "@prisma/client": "^5.11.0"
-  },
-  "devDependencies": {
-    "prisma": "^5.11.0",
-    "tsx": "^3.14.0",
-    "typescript": "^5.4.3"
-  }
-}
