--- example/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["esnext"],
-    "module": "esnext",
-    "target": "es2022",
-
-    "strict": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node",
-    "noEmit": true,
-    "isolatedModules": true,
-    "incremental": true,
-    "noUncheckedIndexedAccess": true
-  },
-  "include": ["index.ts"]
-}
